Day 15 - Coffee Machine
A fully functional coin operated coffee machine 

The machine allows you to order an espresso, latte, and cappuccino. 
The machine allows you to pull a report of the current resources.
The machine allows you to turn the machine off.
The machine allows you to refill the machine resources.
The machine has a charge for each of the coffee's and charges the cost of the coffee against what the user provides as money for it, and if theres less than, the required amount or more the machine replies accordingly.