Day 17 - OOP Quiz
This is an object-oriented quiz game that can be swapped with any set of python dictionary data using the quiz trivia API on https://opentdb.com/api_config.php. There is a set of 10 T/F questions that is asked to the user and then the score is tracked based on the number of correct and incorrect answers they enter, then told at the end of the game.